% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turning_patterns.R
\name{save_image}
\alias{save_image}
\title{Save a Turing Pattern image}
\usage{
save_image(
  tp_grid,
  file_name,
  color_vals = hcl.colors(12, "YlOrRd", rev = TRUE)
)
}
\arguments{
\item{tp_grid}{output turing pattern grid, generated by turing_pattern()}

\item{color_vals}{vector of color values for image display}
}
\description{
save a turing pattern image to a png
}
