% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turning_patterns.R
\name{show_image}
\alias{show_image}
\title{Display a Turing Pattern image}
\usage{
show_image(tp_grid, color_vals = hcl.colors(12, "YlOrRd", rev = TRUE))
}
\arguments{
\item{tp_grid}{output turing pattern grid, generated by turing_pattern()}

\item{color_vals}{vector of color values for image display}
}
\description{
display a turing pattern image
}
